Set MAC address while inserting wifi driver
Restarting the network always reset the MAC address of the wlan0
interface. By default, the MAC address of the wifi driver is set to
00:11:22:33:44:55.

This commit ensures that the MAC address stored in the OTP
corresponds to the default MAC address of the wifi driver by reading
the OTP and passing it as an argument to insmod.

Signed-off-by: Francois Berder <Francois.Berder@imgtec.com>