README: simplify docs
We have moved a lot of documentation to our docs website but not
removed it from here so it is now duplicated and often out of sync.
Therefore have removed most of it except a minimal quickstart guide.

Have also moved contribution section to dedicated file as is
standard meaning it will be picked up and show to any new
conributors by github.

Signed-off-by: Ian Pozella <Ian.Pozella@imgtec.com>