pistachio: Set bus width of sdhost controller to 1
The sdhost controller can support 4 bit mode only if the upper 4 bits
are pulled high. Since these gpios are used on mikrobus connectors, this
constraint is not always met, causing communication failures with the
sdcard when any of these gpios is low.
It has been found that reducing the bus width to 1 avoids the issue
at the expense of performance.

Signed-off-by: Francois Berder <francois.berder@imgtec.com>