Jenkinsfile: don't modify sdk feeds.conf.default
In an effort to create reproducible builds we let the build system
manage cloning feed repos and modify feeds.conf.default to point to
these local feed repos. The issue is that doing this means this
modified config file ends up in the SDK which breaks it. Seeing as
the config file is only used by scripts/feeds we can make the
changes to the config, install required feeds and then revert the
config back.

Signed-off-by: Ian Pozella <Ian.Pozella@imgtec.com>