Jenkinsfile: only index params through their array
For convenience build params are set as top level vars too but this
is not the case on the very first build (which happens a lot due to
the way PRs are handled in the organisation plugin). Therefore make
sure they are used properly.

Secondly make sure that double quotes are used where required for
groovy and shell expansion.

Signed-off-by: Ian Pozella <Ian.Pozella@imgtec.com>